+++ client/src/components/App.jsx
@@ -12,8 +12,7 @@ import {
   MenuDivider,
   Button,
   Flex,
-  extendTheme,
- ChakraProvider
+  extendTheme
 } from '@chakra-ui/react';
 import { ChevronDownIcon } from '@chakra-ui/icons';
 import { Link } from "react-router-dom";
@@ -23,18 +22,6 @@ import axios from 'axios';
 const App = () => {
 
   const [search, setSearch] = useState(null)
-
-  const theme = extendTheme({
-    colors: {
-      brand: {
-        100: "#2E2F30",  //black {header}
-        200: "#8DD8E0",  //blue {border color}
-        300: "#E3444B",  //red  {buttons}
-        400: "#EC7C71",  //orange {button border}
-        500: "#FBFAFA",  //white {subheaders, text}
-      },
-    },
-  })
 
   const searchEvents = (search) => {
     axios.get('/api/events/s', { params: { search: search }})
@@ -50,7 +37,7 @@ const App = () => {
     <div>
       <Flex style={{ borderBottom: '5px solid black' }}>
         <Link to="/">
-          <Heading as='h1' fontSize='6vh' h='11.5vh' color='brand.500' p={8}>bored tour</Heading>
+          <Heading as='h1' fontSize='6vh' h='11.5vh' color='brand.100' p={8}>bored tour</Heading>
         </Link>
         <Spacer />
         <Box p={4}>
