+++ client/src/components/events.js
@@ -1,20 +1,9 @@
-import { Box, Flex, Heading, extendTheme, ChakraProvider } from '@chakra-ui/react';
+import { Box, Flex, Heading, extendTheme } from '@chakra-ui/react';
 import React from 'react';
 import { Header } from './header';
 
 
 const Events = () => {
-  const theme = extendTheme({
-    colors: {
-      brand: {
-        100: "#2E2F30",  //black {header}
-        200: "#8DD8E0",  //blue {border color}
-        300: "#E3444B",  //red  {buttons}
-        400: "#EC7C71",  //orange {button border}
-        500: "#FBFAFA",  //white {subheaders, text}
-      },
-    },
-  })
 
   return (
   <Box>
