+++ client/src/components/header.js
@@ -10,8 +10,7 @@ import {
   MenuDivider,
   Button,
   Flex,
-  extendTheme,
- ChakraProvider
+  extendTheme
 } from '@chakra-ui/react';
 import { ChevronDownIcon } from '@chakra-ui/icons';
 import { Link } from "react-router-dom";
@@ -19,17 +18,6 @@ import { Link } from "react-router-dom";
 
 export const Header = () => {
 
-  const theme = extendTheme({
-    colors: {
-      brand: {
-        100: "#2E2F30",  //black {header}
-        200: "#8DD8E0",  //blue {border color}
-        300: "#E3444B",  //red  {buttons}
-        400: "#EC7C71",  //orange {button border}
-        500: "#FBFAFA",  //white {subheaders, text}
-      },
-    },
-  })
 
  return(
     <Box
