+++ client/src/components/homepage/event.jsx
@@ -1,4 +1,4 @@
-import { Heading, Container, Box, Spacer, Flex, Button, ButtonGroup, FormControl, extendTheme, ChakraProvider } from '@chakra-ui/react';
+import { Heading, Container, Box, Spacer, Flex, Button, ButtonGroup, FormControl } from '@chakra-ui/react';
 import { ChevronLeftIcon, ChevronRightIcon } from '@chakra-ui/icons'
 import React, { useRef, useEffect, useState } from 'react';
