+++ client/src/components/map.js
@@ -1,8 +1,9 @@
-import { Flex, extendTheme, ChakraProvider } from '@chakra-ui/react';
+import { Flex, extendTheme } from '@chakra-ui/react';
 import React, { useState, useEffect } from 'react';
 import { GoogleMap, LoadScript, Marker } from '@react-google-maps/api';
 import axios from 'axios';
 import Geocode from "react-geocode";
+import firebaseConfig from '../../../config';
 
 const Map = ({ address }) => {
   const mapStyles = {
@@ -11,22 +12,10 @@ const Map = ({ address }) => {
     float: "right"
   };
 
-  const theme = extendTheme({
-    colors: {
-      brand: {
-        100: "#2E2F30",  //black {header}
-        200: "#8DD8E0",  //blue {border color}
-        300: "#E3444B",  //red  {buttons}
-        400: "#EC7C71",  //orange {button border}
-        500: "#FBFAFA",  //white {subheaders, text}
-      },
-    },
-  })
-
 
   const [location, setLocation] = useState({});
 
-  Geocode.setApiKey(process.env.GOOGLE_API);
+  Geocode.setApiKey(firebaseConfig.GOOGLE_API);
   useEffect(() => {
     Geocode.fromAddress(address).then(
       (response) => {
@@ -40,7 +29,7 @@ const Map = ({ address }) => {
  };
   return (
     <LoadScript
-      googleMapsApiKey={process.env.GOOGLE_API}>
+      googleMapsApiKey={firebaseConfig.GOOGLE_API}>
       <GoogleMap
           mapContainerStyle={mapStyles}
           zoom={12}
