+++ client/src/components/notfound.js
@@ -1,4 +1,4 @@
-import { Box, Flex, Heading, Button, Spacer, extendTheme, ChakraProvider } from '@chakra-ui/react';
+import { Box, Flex, Heading, Button, Spacer, extendTheme } from '@chakra-ui/react';
 import React from 'react';
 import { Header } from './header';
 import { Link } from 'react-router-dom';
@@ -6,17 +6,6 @@ import { Link } from 'react-router-dom';
 
 const NotFound = () => {
 
-  const theme = extendTheme({
-    colors: {
-      brand: {
-        100: "#2E2F30",  //black {header}
-        200: "#8DD8E0",  //blue {border color}
-        300: "#E3444B",  //red  {buttons}
-        400: "#EC7C71",  //orange {button border}
-        500: "#FBFAFA",  //white {subheaders, text}
-      },
-    },
-  })
 
   return (
   <Box>
