+++ client/src/index.js
@@ -12,9 +12,11 @@ import { Register } from './components/register.jsx';
 import Preferences from './components/preferences/Preferences.jsx';
 import EventDetail from './components/eventDetail';
 
+import theme from './theme.js';
+
 render(
   <BrowserRouter>
-  <ChakraProvider>
+  <ChakraProvider theme={theme}>
   <Routes>
       <Route path="/" element={<App />} />
       <Route path="friends" element={<Friends />} />
